+++ index.js
@@ -362,6 +362,7 @@ export * from './src/components/attrsHelpers.js';
 export * from './src/components/modelHelper.js';
 export * from './src/components/composables.js';
 export {
+  legendSymbol,
   LegendType,
   StyleRowType,
   getLegendEntries,